% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{Estimation_Function}
\alias{Estimation_Function}
\title{Estimate}
\usage{
Estimation_Function(Y, AX, LAB, k, ks, lambda = 0.1)
}
\arguments{
\item{Y}{Matrix of the output variable rows=length of time series, cols=number of individuals}

\item{AX}{Matrix of the input variable rows=length of time series, cols=number of individuals x number of explanatory variables}

\item{LAB}{vector of group membership of each individual}

\item{k}{number of?}

\item{ks}{vector of number of group specific factors}

\item{lambda}{regularization parameter}
}
\value{
A list containing PredL, PredXB, PredG and Bes
}
\description{
This function does the following
}
